+++ examples/examples.cmake
@@ -1,3 +1,0 @@
-include(${CMAKE_CURRENT_LIST_DIR}/simple_lut/simple_lut.cmake)
-include(${CMAKE_CURRENT_LIST_DIR}/simple_sdm/simple_sdm.cmake)
-include(${CMAKE_CURRENT_LIST_DIR}/i2s_slave_lut/i2s_slave_lut.cmake)