+++ tests/test_deps.cmake
@@ -1,0 +1,2 @@
+set(APP_DEPENDENT_MODULES "lib_sw_pll"
+    )