+++ CHANGELOG.rst
@@ -1,6 +1,16 @@
 lib_sw_pll change log
 =====================
 
+2.3.0
+-----
+
+  * CHANGED: Updated pll_calc script to separate out integer and fractional
+    divider values for easier user interpretation
+  * CHANGED: Improved PLL solution selection in app_pll_model to ensure
+    sufficient positive and negative range from nominal frequency
+  * CHANGED: Uses Xcommon Cmake instead of custom Cmake
+  * FIXED: Python models and test type error on later numpy versions
+
 2.2.0
 -----